code refactor - moved loginService, orderEditService and orderSearchService files to services folder
