+++ src/client/components/clinicPortal/facility/clinicFacilityGrid/addFacility.jsx
@@ -1,17 +1,5 @@
 import React, { Component } from "react";
-import { Modal, Button } from "react-bootstrap";
-import {
-	saveOrderEditData,
-	updateResultPDF,
-} from "../../../../clinicPortalServices/orderEditService";
-import moment from "moment";
-import {
-	faxTypes,
-	results,
-} from "../../patientSearch/clinicPatientGrid/optionsData";
-import { testTypes } from "../../patientSearch/clinicPatientGrid/optionsData";
-import { fetchOrderFaxData } from "../../../../clinicPortalServices/orderEditService";
-import { Tooltip, OverlayTrigger } from "react-bootstrap";
+import { Modal } from "react-bootstrap";
 import FacilityDetails from "./facilityDetails";
 
 export default class AddFacility extends Component {
@@ -20,7 +8,6 @@ export default class AddFacility extends Component {
 		//console.log(props);
 		this.state = {
 			show: false,
-			//refreshGrid: props.data.refreshGrid,
 			firstName: "",
 			lastName: "",
 			code: "",
@@ -39,7 +26,6 @@ export default class AddFacility extends Component {
 	handleClose = () => {
 		this.setState({ show: false });
 	};
-	
 
 	render() {
 		return (
@@ -50,7 +36,7 @@ export default class AddFacility extends Component {
 				>
 					<i class="fas fa-user-plus"></i> Add Facility
 				</button>
-                <Modal
+				<Modal
 					size="lg"
 					aria-labelledby="contained-modal-title-vcenter"
 					centered
@@ -61,9 +47,7 @@ export default class AddFacility extends Component {
 						<Modal.Title>Add Facility Information</Modal.Title>
 					</Modal.Header>
 					<Modal.Body>
-						<FacilityDetails
-							handleClose={this.handleClose}
-						/>
+						<FacilityDetails handleClose={this.handleClose} />
 					</Modal.Body>
 				</Modal>
 			</div>
