+++ src/client/services/clinicPortalServices/loginService.js
@@ -1,5 +1,6 @@
-import { serviceConstants } from "../patientPortalServices/constants";
-import { getUserAuthToken } from "../utils/util";
+import { serviceConstants } from "../../patientPortalServices/constants";
+import { getUserAuthToken } from "../../utils/util";
+
 export const authenticateAndFetchUserDetails = (username, password) => {
 	return fetch(`${serviceConstants.API_HOST_NAME}/user/v1/authenticate`, {
 		method: "POST",
@@ -7,8 +8,8 @@ export const authenticateAndFetchUserDetails = (username, password) => {
 			"Content-Type": "application/json",
 		},
 		body: JSON.stringify({
-			user_name:username,
-			password:password
+			user_name: username,
+			password: password,
 			//user_name: "test1",
 			//password: "Test#123",
 		}),
@@ -27,12 +28,12 @@ export const authenticateAndFetchUserDetails = (username, password) => {
 
 export const logout = () => {
 	//console.log("calling logout api");
-	var token  = getUserAuthToken();
+	var token = getUserAuthToken();
 	return fetch(`${serviceConstants.API_HOST_NAME}/user/v1/logout`, {
 		method: "POST",
 		headers: {
 			"Content-Type": "application/json",
-			"Authorization" : "Bearer " + token
-		}
+			Authorization: "Bearer " + token,
+		},
 	}).then((response) => response.json());
 };
